+++ src/ExcalidrawEditor.ts
@@ -1,312 +1,140 @@
 import * as vscode from "vscode";
 import * as fs from "fs";
-import * as path from "path";
+import { resolve } from "path";
 
-export class ExcalidrawEditorProvider implements vscode.CustomTextEditorProvider {
-	public static register(context: vscode.ExtensionContext): vscode.Disposable {
-		vscode.commands.registerCommand("excalidraw.export.svg", () => {
-			ExcalidrawEditorProvider.exportToIMG("svg");
-		});
-		vscode.commands.registerCommand("excalidraw.export.png", () => {
-			ExcalidrawEditorProvider.exportToIMG("png");
-		});
+export class ExcalidrawTextEditorProvider
+  implements vscode.CustomTextEditorProvider {
+  public static register(context: vscode.ExtensionContext): vscode.Disposable {
+    const provider = new ExcalidrawTextEditorProvider(context);
+    const providerRegistration = vscode.window.registerCustomEditorProvider(
+      ExcalidrawTextEditorProvider.viewType,
+      provider
+    );
+    return providerRegistration;
+  }
 
-		const provider = new ExcalidrawEditorProvider(context);
-		const providerRegistration = vscode.window.registerCustomEditorProvider(
-			ExcalidrawEditorProvider.viewType,
-			provider
-		);
-		return providerRegistration;
-	}
+  private static readonly viewType = "editor.excalidraw";
 
-	private static readonly viewType = "editor.excalidraw";
-	public static exportToIMG: Function = () => {
-		vscode.window.showErrorMessage(
-			"At least one excalidraw editor must be active to use this command!"
-		);
-	};
+  static activeEditor: ExcalidrawEditor | undefined;
 
-	constructor(private readonly context: vscode.ExtensionContext) { }
 
-	/**
-	 * Called when our custom editor is opened.
-	 *
-	 *
-	 */
-	public async resolveCustomTextEditor(
-		document: vscode.TextDocument,
-		webviewPanel: vscode.WebviewPanel,
-		_token: vscode.CancellationToken
-	): Promise<void> {
-		// Setup initial content for the webview
-		webviewPanel.webview.options = {
-			enableScripts: true,
-		};
-		let excalidrawConfig = vscode.workspace.getConfiguration("excalidraw")
-		webviewPanel.webview.html = this.getHtmlForWebview(document, this.getInitialData(document, excalidrawConfig.get("theme", "auto")));
+  constructor(private readonly context: vscode.ExtensionContext) { }
 
-		const refreshTheme = () => {
-			const theme = excalidrawConfig
-				.get("theme", "auto");
-			webviewPanel.webview.postMessage({
-				type: "refresh-theme",
-				theme: theme,
-			});
-		};
-		const changeConfigurationSubscription = vscode.workspace.onDidChangeConfiguration(
-			(e) => {
-				if (e.affectsConfiguration("excalidraw")) {
-					excalidrawConfig = vscode.workspace.getConfiguration("excalidraw")
-					refreshTheme();
-				}
-			}
-		);
+  public async resolveCustomTextEditor(
+    document: vscode.TextDocument,
+    webviewPanel: vscode.WebviewPanel,
+    _token: vscode.CancellationToken
+  ): Promise<void> {
 
-		const exportToIMG = (extension: string) => {
-			const exportConfig = vscode.workspace.getConfiguration(
-				"excalidraw.export"
-			);
-			this.getExportFilename(document, extension).then((uri) => {
-				if (uri !== undefined)
-					webviewPanel.webview.postMessage({
-						type: `export-to-${extension}`,
-						path: uri.fsPath,
-						exportConfig: {
-							exportBackground: exportConfig.get("exportBackground"),
-							shouldAddWatermark: exportConfig.get("shouldAddWatermark"),
-							exportWithDarkMode: exportConfig.get("exportWithDarkMode"),
-							exportEmbedScene: exportConfig.get("exportEmbedScene"),
-						},
-					});
-			});
-		};
+    const editor = new ExcalidrawEditor(webviewPanel, this.context);
+    editor.edit(document);
 
-		ExcalidrawEditorProvider.exportToIMG = exportToIMG;
-		webviewPanel.onDidChangeViewState((e) => {
-			if (document.uri.scheme === "git") return;
-			if (e.webviewPanel.active) {
-				ExcalidrawEditorProvider.exportToIMG = exportToIMG;
-				refreshTheme();
-				vscode.commands.executeCommand(
-					"setContext",
-					"excalidraw.focused",
-					true
-				);
-			} else {
-				vscode.commands.executeCommand(
-					"setContext",
-					"excalidraw.focused",
-					false
-				);
-			}
-		});
+    const onDidChangeViewState = webviewPanel.onDidChangeViewState((e) => {
+      ExcalidrawTextEditorProvider.activeEditor = e.webviewPanel.active ? editor : undefined;
+    })
 
-		const updateWebview = () => {
-			const { elements, appState } = this.getInitialData(document, excalidrawConfig.get("theme", "auto"));
-			webviewPanel.webview.postMessage({
-				type: "update",
-				elements: elements,
-				appState: appState,
-			});
-		};
+    webviewPanel.onDidDispose(
+      onDidChangeViewState.dispose
+    )
 
-		const changeDocumentSubscription = vscode.workspace.onDidChangeTextDocument(
-			(e) => {
-				if (
-					e.document.uri.toString() === document.uri.toString() &&
-					e.contentChanges.length > 0
-				) {
-					updateWebview();
-				}
-			}
-		);
-
-		// Make sure we get rid of the listener when our editor is closed.
-		webviewPanel.onDidDispose(() => {
-			changeDocumentSubscription.dispose();
-			changeConfigurationSubscription.dispose();
-		});
-
-		// Receive message from the webview.
-		webviewPanel.webview.onDidReceiveMessage((e) => {
-			switch (e.type) {
-				case "init":
-					vscode.commands.executeCommand(
-						"setContext",
-						"excalidraw.focused",
-						true
-					);
-					return;
-				case "library":
-					this.context.globalState.update("libraryItems", e.items)
-					break;
-				case "update":
-					this.updateTextDocument(document, e.elements, e.appState).then(() => {
-						if (excalidrawConfig.get("autoSave", true))
-							document.save()
-					}
-					)
-					return;
-				case "svg-export":
-					createDirIfNeeded(e.path)
-					console.log(e)
-					fs.writeFile(e.path, e.svg, (err) => {
-						if (err) vscode.window.showErrorMessage(err.message);
-						else vscode.window.showInformationMessage(`Export Successful`, "Open").then(msg => {
-							if (msg === "Open")
-								vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(e.path))
-						});
-					});
-					return;
-				case "png-export":
-					createDirIfNeeded(e.path)
-					var data = e.png.replace(/^data:image\/png;base64,/, "");
-					var buf = Buffer.from(data, "base64");
-					fs.writeFile(e.path, buf, (err) => {
-						if (err) vscode.window.showErrorMessage(err.message);
-						else vscode.window.showInformationMessage(`Export Successful`, "Open").then(msg => {
-							if (msg === "Open")
-								vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(e.path))
-						});
-					});
-					return;
-				case "refresh-theme":
-					const theme = vscode.workspace
-						.getConfiguration("excalidraw")
-						.get("theme", "auto");
-					webviewPanel.webview.postMessage({
-						type: "refresh-theme",
-						theme: theme,
-					});
-				case "log":
-					console.log(e.msg);
-					return;
-			}
-		});
-	}
-
-	/**
-	 * Get the static html used for the editor webviews.
-	 */
-	private getHtmlForWebview(document: vscode.TextDocument, initialData: any): string {
-		const htmlFile = vscode.Uri.joinPath(
-			this.context.extensionUri,
-			"media",
-			"index.html"
-		);
-		let content = fs.readFileSync(htmlFile.fsPath, "utf8");
-		if (initialData.readOnly)
-			content = content.replace(
-				/(<style>)(<\/style>)/,
-				`$1
-			.Island {
-				display: none !important;
-			}
-			$2`
-			);
-
-		content = content.replace(
-			/\$\{initialData\}/,
-			`<script>window.initialData = ${JSON.stringify(initialData)}</script>`
-		);
-
-		return content;
-	}
-
-	/**
-	 * Try to get a current document as json text.
-	 */
-	private getInitialData(document: vscode.TextDocument, theme: string): any {
-		const text = document.getText();
-		const libraryItems = this.context.globalState.get("libraryItems", [])
-
-		if (text.trim().length === 0) {
-			return { elements: [], appState: {}, themeConfig: theme, libraryItems: libraryItems };
-		}
-
-		try {
-			const json = JSON.parse(text);
-			const { elements, appState } = json;
-			const initialData = {
-				elements: elements,
-				appState: { ...appState },
-				libraryItems: libraryItems,
-				themeConfig: theme,
-				readOnly: document.uri.scheme === "git"
-			};
-			return initialData;
-		} catch {
-			throw new Error(
-				"Could not get document as json. Content is not valid json"
-			);
-		}
-	}
-
-	private getExportFilename(
-		document: vscode.TextDocument,
-		extension: string
-	): Thenable<vscode.Uri | undefined> {
-		const dirname = path.dirname(document.uri.fsPath);
-		const basename = path.basename(
-			document.uri.fsPath,
-			path.extname(document.uri.fsPath)
-		);
-		const filePath = path.join(dirname, `${basename}.${extension}`);
-		return vscode.window.showSaveDialog({
-			defaultUri: vscode.Uri.file(filePath),
-			filters: { Images: [extension] },
-		});
-	}
-
-	/**
-	 * Write out the json to a given document.
-	 */
-	private updateTextDocument(
-		document: vscode.TextDocument,
-		elements: Record<string, unknown>,
-		appState: Record<string, unknown>
-	): Thenable<boolean> {
-		const newContent = JSON.stringify(
-			{
-				type: "excalidraw",
-				version: 2,
-				source: "https://excalidraw.com",
-				elements: elements,
-				appState: appState,
-			},
-			null,
-			2
-		);
-
-		const edit = new vscode.WorkspaceEdit();
-
-		edit.replace(
-			document.uri,
-			new vscode.Range(0, 0, document.lineCount, 0),
-			newContent
-		);
-
-		return vscode.workspace.applyEdit(edit);
-	}
+    ExcalidrawTextEditorProvider.activeEditor = editor;
+  }
 }
 
-export const debounce = <T extends (...args: any[]) => any>(
-	callback: T,
-	waitFor: number
-) => {
-	let timeout: ReturnType<typeof setTimeout>;
-	return (...args: Parameters<T>): ReturnType<T> => {
-		let result: any;
-		timeout && clearTimeout(timeout);
-		timeout = setTimeout(() => {
-			result = callback(...args);
-		}, waitFor);
-		return result;
-	};
-};
+class ExcalidrawEditor {
+  private config: vscode.WorkspaceConfiguration;
+  constructor(
+    readonly webviewPanel: vscode.WebviewPanel, private readonly context: vscode.ExtensionContext) {
 
-function createDirIfNeeded(filepath: string) {
-	const dirname = path.dirname(filepath)
-	if (!fs.existsSync(dirname))
-		fs.mkdirSync(dirname, { recursive: true })
+    webviewPanel.webview.options = {
+      enableScripts: true,
+    };
+
+    this.config = vscode.workspace.getConfiguration("excalidraw");
+  }
+
+  public edit(document: vscode.TextDocument) {
+    // Setup initial content for the webview
+    // Receive message from the webview.
+    const onDidReceiveMessage = this.webviewPanel.webview.onDidReceiveMessage((msg) => {
+      switch (msg.type) {
+        case "library-change":
+          this.context.globalState.update("libraryItems", msg.libraryItems);
+          break;
+        case "change":
+          this.updateTextDocument(document, msg.content).then(() => {
+            if (this.config.get("autoSave")) document.save();
+          });
+          return;
+        case "log":
+          console.log(msg.msg);
+          return;
+      }
+    });
+
+    this.webviewPanel.webview.html = this.getHtmlForWebview(
+      {
+        content: document.getText(),
+        contentType: document.fileName.endsWith(".svg") ? "image/svg+xml" : "application/json",
+        libraryItems: this.context.globalState.get("libraryItems") || [],
+      }
+    );
+
+    this.webviewPanel.onDidDispose(
+      onDidReceiveMessage.dispose
+    )
+
+  }
+  /**
+* Apply Edit on Document
+*/
+  private updateTextDocument(
+    document: vscode.TextDocument,
+    content: string
+  ): Thenable<boolean> {
+    const edit = new vscode.WorkspaceEdit();
+
+    edit.replace(
+      document.uri,
+      new vscode.Range(0, 0, document.lineCount, 0),
+      content
+    );
+
+    return vscode.workspace.applyEdit(edit);
+  }
+
+  public importLibrary(libraryUrl: string, csrfToken: string) {
+    this.webviewPanel.webview.postMessage({ type: "import-library", libraryUrl, csrfToken });
+  }
+
+  private getHtmlForWebview(
+    data: Record<string, unknown>
+  ): string {
+    const htmlFile = vscode.Uri.file(resolve(this.context.extensionPath, "media", "index.html"));
+    let html = fs.readFileSync(htmlFile.fsPath, "utf8");
+
+    // Fix resources path
+    html = html.replace(
+      /(<link.+?href="|<script.+?src="|<img.+?src="|url\(")(.+?)"/g,
+      (m, $1, $2) => {
+        const resourcePath = `${this.context.extensionPath}/media${$2}`;
+        return (
+          $1 +
+          vscode.Uri.from({ path: resourcePath, scheme: "vscode-resource" })
+            .toString() +
+          '"'
+        );
+      }
+    );
+
+    const base64Config = Buffer.from(JSON.stringify(data), "utf-8").toString("base64");
+
+    // Pass document uri to the webview
+    html = html.replace(
+      "{data-excalidraw}",
+      base64Config
+    );
+
+    return html;
+  }
 }